// ADL/bbog-bv-ionic/src/new-app/modules/dashboard/models/dashboard-cardActication.model.ts
export interface DashboardCardActivation {
    cardsActivation?: any;
    CardFirstText?: string;
    CardSecondText?: string;
    creditCardActivation?: boolean;
    isActivationCard?: boolean;
    multipleCardsActivation?: boolean;
    debitCardActivation?: boolean;
    activationCardSelected?: any;
}